feat: refactors findAllResults methods for extending entities on publishedProjects
